requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.1
fastapi==0.104.1
uvicorn==0.24.0 